koord-manager: refactor mid resource calculate logic.
